_Tex class introduced to seperate responsibility
