+++ manim.cfg
@@ -1,4 +1,3 @@
 [CLI]
 background_color = WHITE
 frame_width = 6
-frozen_frame = False
